GPSService Android App
First Commit - GPS Tracker creates a service which in turn will
implement Location Listener (Android API to use GPS), and since it’s a
service, it will keep on running in a background. So irrespective of any
GPS app is running or not, GPS will keep on searching for satellite even
when you close any GPS app.